fix: Fixed cache entries never being cleaned
